Add files via upload
Problem: 
The company is trying to retain its customers using their purchase history and transaction details and find which type of life stage customers spend mostly on a specific brand.

Importance: 
Customer Analytics is a particularly important sphere in today’s scenario, and a lot of investment and improvements are being done in order to achieve maximum output. This research helps the companies to retain its existing customers and provide them with their preferred product and particularly increase their business by targeting a certain set of customers.

Value Addition: 
We will be presenting a strategic recommendation to the client that is supported by data which she can then use for the upcoming category review however to do so we will analyze the data to understand the current purchasing trends and behaviors. The client is particularly interested in customer segments and their snacks purchasing behavior. We will consider different metrics which would help describe the customers’ purchasing behavior.